third_party/WebKit/LayoutTests/fast/viewport/viewport-52-expected.txt

CONSOLE ERROR: line 3: The value "garbage" for key "width" is invalid, and has been ignored.
ALERT: viewport size 64x70.4 scale 5 with limits [5, 5] and userScalable true
